feat: added a configuration file and started to work on transforming the data

+++ src/etl_pipeline.py
@@ -1,9 +1,9 @@
 import json
 import logging
 import os
-from multiprocessing import Value
 from pathlib import Path
 
+import pandas as pd
 import requests
 from dotenv import load_dotenv
 
@@ -33,14 +33,35 @@ def extract_products():
 
 
 def save_raw_data(data):
+    """From the extracted data we save this to a file"""
     if not data:
         raise ValueError("No data to save")
+
     if not RAW_DATA_DIRECTORY.exists():
         RAW_DATA_DIRECTORY.mkdir(parents=True)
+
     file_path = RAW_DATA_DIRECTORY / "products.json"
+    if file_path.exists():
+        logging.info(
+            f"Raw data already exists at {file_path}. Please delete it before saving new data."
+        )
+        return
     with open(file_path, "w") as file:
         json.dump(data, file)
     logging.info(f"Raw data saved to {file_path}")
+
+
+def transform_data(data):
+    """Transform the raw data into a pandas DataFrame"""
+    df = pd.DataFrame(data)
+    # extract rating fields (average score and review count) into separate columns
+    df["customer_rating"] = df["rating"].apply(lambda x: x["rate"])  # rating_mean
+    df["customer_reviews"] = df["rating"].apply(lambda x: x["count"])  # rating_samples
+    df = df.drop(columns=["rating"], axis=1)
+
+    # Renaming columns
+    df = df.rename(columns={"image": "image_url"})
+    return df
 
 
 data = extract_products()
